(maint) Prepare cert bundle and keystore prior to install
Previously, we would create our cert bundle during build time
when we installed it and potentially also create a Java keystore.

This is error prone and the formats of the cert bundle and keystore are
platform agnostic so we can build them ahead of time (ideally when
updating the certs from upstream).

This patch adds a target to the Makefile, `prepare` which will create
both a cert bundle and a keystore. This patch also refactors the
Makefile so that the `install` target will install the keystore in
addition to the bundle and no longer depends on the bundle creation
target. Now `install` is the default make target and the puppet-runtime
project needs only call `make` during its build.